+++ src/main/java/org/truffleruby/platform/Platform.java
@@ -31,18 +31,19 @@ import org.truffleruby.shared.BasicPlatform;
 
 public abstract class Platform extends BasicPlatform {
 
-    public static final String LIBSUFFIX = determineLibExt();
+    public static final String LIB_SUFFIX = determineLibSuffix();
+    public static final String CEXT_SUFFIX = OS == OS_TYPE.DARWIN ? ".bundle" : LIB_SUFFIX;
 
     public static final boolean IS_WINDOWS = OS == OS_TYPE.WINDOWS;
 
-    private static final String determineLibExt() {
+    private static String determineLibSuffix() {
         switch (OS) {
             case WINDOWS:
-                return "dll";
+                return ".dll";
             case DARWIN:
-                return "dylib";
+                return ".dylib";
             default:
-                return "so";
+                return ".so";
         }
     }