+++ CHANGELOG.md
@@ -53,6 +53,7 @@ Compatibility:
 * JCodings has been updated from 1.0.42 to 1.0.45.
 * Joni has been updated from 2.1.25 to 2.1.30.
 * Implemented `Method#<<` and `Method#>>` (#1821).
+* The `.bundle` file extension is now used for C extensions on macOS (#1819, #1837).
 
 Performance: